+++ api/app/controllers/spree/api/checkouts_controller.rb
@@ -52,8 +52,8 @@ module Spree
           # respond_to check is necessary due to issue described in #2910
           object_params = nested_params
           if @order.has_checkout_step?("payment") && @order.payment?
-            if object_params[:payment_source].present? && source_params = object_params.delete(:payment_source)[object_params[:order][:payments_attributes].first[:payment_method_id].underscore]
-              object_params[:order][:payments_attributes].first[:source_attributes] = source_params
+            if object_params[:payment_source].present? && source_params = object_params.delete(:payment_source)[object_params[:payments_attributes].first[:payment_method_id].underscore]
+              object_params[:payments_attributes].first[:source_attributes] = source_params
             end
             if object_params[:payments_attributes]
               object_params[:payments_attributes].first[:amount] = @order.total.to_s